+++ src/Bridge/StorageFacade.php
@@ -13,6 +13,9 @@ use Heptacom\HeptaConnect\Storage\Base\Contract\Action\Identity\IdentityOverview
 use Heptacom\HeptaConnect\Storage\Base\Contract\Action\Identity\IdentityPersistActionInterface;
 use Heptacom\HeptaConnect\Storage\Base\Contract\Action\Identity\IdentityReflectActionInterface;
 use Heptacom\HeptaConnect\Storage\Base\Contract\Action\IdentityError\IdentityErrorCreateActionInterface;
+use Heptacom\HeptaConnect\Storage\Base\Contract\Action\IdentityRedirect\IdentityRedirectCreateActionInterface;
+use Heptacom\HeptaConnect\Storage\Base\Contract\Action\IdentityRedirect\IdentityRedirectDeleteActionInterface;
+use Heptacom\HeptaConnect\Storage\Base\Contract\Action\IdentityRedirect\IdentityRedirectOverviewActionInterface;
 use Heptacom\HeptaConnect\Storage\Base\Contract\Action\Job\JobCreateActionInterface;
 use Heptacom\HeptaConnect\Storage\Base\Contract\Action\Job\JobDeleteActionInterface;
 use Heptacom\HeptaConnect\Storage\Base\Contract\Action\Job\JobFailActionInterface;
@@ -61,6 +64,9 @@ use Heptacom\HeptaConnect\Storage\ShopwareDal\Action\Identity\IdentityOverview;
 use Heptacom\HeptaConnect\Storage\ShopwareDal\Action\Identity\IdentityPersist;
 use Heptacom\HeptaConnect\Storage\ShopwareDal\Action\Identity\IdentityReflect;
 use Heptacom\HeptaConnect\Storage\ShopwareDal\Action\IdentityError\IdentityErrorCreate;
+use Heptacom\HeptaConnect\Storage\ShopwareDal\Action\IdentityRedirect\IdentityRedirectCreate;
+use Heptacom\HeptaConnect\Storage\ShopwareDal\Action\IdentityRedirect\IdentityRedirectDelete;
+use Heptacom\HeptaConnect\Storage\ShopwareDal\Action\IdentityRedirect\IdentityRedirectOverview;
 use Heptacom\HeptaConnect\Storage\ShopwareDal\Action\Job\JobCreate;
 use Heptacom\HeptaConnect\Storage\ShopwareDal\Action\Job\JobDelete;
 use Heptacom\HeptaConnect\Storage\ShopwareDal\Action\Job\JobFail;
@@ -140,6 +146,25 @@ class StorageFacade extends AbstractSingletonStorageFacade
     protected function createFileReferenceGetRequestAction(): FileReferenceGetRequestActionInterface
     {
         return new FileReferenceGetRequestAction($this->getQueryFactory());
+    }
+
+    protected function createIdentityRedirectCreateActionInterface(): IdentityRedirectCreateActionInterface
+    {
+        return new IdentityRedirectCreate(
+            $this->connection,
+            $this->getStorageKeyGenerator(),
+            $this->getEntityTypeAccessor()
+        );
+    }
+
+    protected function createIdentityRedirectDeleteActionInterface(): IdentityRedirectDeleteActionInterface
+    {
+        return new IdentityRedirectDelete($this->getQueryFactory());
+    }
+
+    protected function createIdentityRedirectOverviewActionInterface(): IdentityRedirectOverviewActionInterface
+    {
+        return new IdentityRedirectOverview($this->getQueryFactory());
     }
 
     protected function createFileReferencePersistRequestAction(): FileReferencePersistRequestActionInterface