+++ src/StorageKey/IdentityRedirectStorageKey.php
@@ -1,0 +1,11 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Heptacom\HeptaConnect\Storage\ShopwareDal\StorageKey;
+
+use Heptacom\HeptaConnect\Storage\Base\Contract\IdentityRedirectKeyInterface;
+
+final class IdentityRedirectStorageKey extends AbstractStorageKey implements IdentityRedirectKeyInterface
+{
+}