+++ test/Suite/Action/WebHttpHandlerConfigurationTest.php
@@ -4,45 +4,34 @@ declare(strict_types=1);
 
 namespace Heptacom\HeptaConnect\Storage\ShopwareDal\Test\Suite\Action;
 
-use Doctrine\DBAL\Connection;
 use Heptacom\HeptaConnect\Storage\Base\Bridge\Contract\StorageFacadeInterface;
 use Heptacom\HeptaConnect\Storage\ShopwareDal\Bridge\StorageFacade;
-use Heptacom\HeptaConnect\TestSuite\Storage\Action\PortalNodeStorageTestContract;
+use Heptacom\HeptaConnect\TestSuite\Storage\Action\WebHttpHandlerConfigurationTestContract;
 
 /**
  * @covers \Heptacom\HeptaConnect\Storage\ShopwareDal\Action\PortalNode\PortalNodeCreate
  * @covers \Heptacom\HeptaConnect\Storage\ShopwareDal\Action\PortalNode\PortalNodeDelete
- * @covers \Heptacom\HeptaConnect\Storage\ShopwareDal\Action\PortalNodeStorage\PortalNodeStorageClear
- * @covers \Heptacom\HeptaConnect\Storage\ShopwareDal\Action\PortalNodeStorage\PortalNodeStorageDelete
- * @covers \Heptacom\HeptaConnect\Storage\ShopwareDal\Action\PortalNodeStorage\PortalNodeStorageGet
- * @covers \Heptacom\HeptaConnect\Storage\ShopwareDal\Action\PortalNodeStorage\PortalNodeStorageList
- * @covers \Heptacom\HeptaConnect\Storage\ShopwareDal\Action\PortalNodeStorage\PortalNodeStorageSet
+ * @covers \Heptacom\HeptaConnect\Storage\ShopwareDal\Action\WebHttpHandlerConfiguration\WebHttpHandlerConfigurationFind
+ * @covers \Heptacom\HeptaConnect\Storage\ShopwareDal\Action\WebHttpHandlerConfiguration\WebHttpHandlerConfigurationSet
  * @covers \Heptacom\HeptaConnect\Storage\ShopwareDal\Bridge\StorageFacade
  * @covers \Heptacom\HeptaConnect\Storage\ShopwareDal\PortalNodeAliasAccessor
+ * @covers \Heptacom\HeptaConnect\Storage\ShopwareDal\StorageKeyGenerator
  * @covers \Heptacom\HeptaConnect\Storage\ShopwareDal\StorageKey\AbstractStorageKey
  * @covers \Heptacom\HeptaConnect\Storage\ShopwareDal\StorageKey\PortalNodeStorageKey
- * @covers \Heptacom\HeptaConnect\Storage\ShopwareDal\StorageKeyGenerator
  * @covers \Heptacom\HeptaConnect\Storage\ShopwareDal\Support\DateTime
  * @covers \Heptacom\HeptaConnect\Storage\ShopwareDal\Support\Id
  * @covers \Heptacom\HeptaConnect\Storage\ShopwareDal\Support\Query\QueryBuilder
  * @covers \Heptacom\HeptaConnect\Storage\ShopwareDal\Support\Query\QueryFactory
  * @covers \Heptacom\HeptaConnect\Storage\ShopwareDal\Support\Query\QueryIterator
  * @covers \Heptacom\HeptaConnect\Storage\ShopwareDal\Test\TestCase
+ * @covers \Heptacom\HeptaConnect\Storage\ShopwareDal\WebHttpHandlerAccessor
+ * @covers \Heptacom\HeptaConnect\Storage\ShopwareDal\WebHttpHandlerPathAccessor
+ * @covers \Heptacom\HeptaConnect\Storage\ShopwareDal\WebHttpHandlerPathIdResolver
  */
-class PortalNodeStorageTest extends PortalNodeStorageTestContract
+class WebHttpHandlerConfigurationTest extends WebHttpHandlerConfigurationTestContract
 {
-    public function testUsageOfPreviewPortalFails(): void
-    {
-        $this->expectNotToPerformDatabaseQueries();
-        parent::testUsageOfPreviewPortalFails();
-    }
-
     protected function createStorageFacade(): StorageFacadeInterface
     {
-        $kernel = $this->kernel;
-        /** @var Connection $connection */
-        $connection = $kernel->getContainer()->get(Connection::class);
-
-        return new StorageFacade($connection);
+        return new StorageFacade($this->getConnection());
     }
 }